Add template for Klaw landing page
This adds a custom template for the landing
page that is used to link to the actual docs.

The TailwindCSS play CDN was used for the landing
page template to quickly add minimal styles for
further development. The play CDN should however
not be used in production, eventually.

Another consideration is that although we probably
don't want to include all of the custom CSS for the
landing page in the HTML head, loading them from the
static assets has some problems because
`sphinx-autobuild` was not able to regenerate those
assets on the fly, making the hot-reload useless
for the static assets.

The styles of the header, footer etc. can later
be applied to the rest of the docs as we switch
the Sphinx theme.

[STORM-412]