Update docs theme
This changes the base theme for the
docs from the Sphinx RTD theme to the
`Furo` theme that's also in use in our
dev portal.

The theming is not yet pixel perfect
with the Figma design but the overall
colors etc. should be now more or less
matched.

TailwindCSS is also installed as an NPM
package in this commit as Tailwind does
not suggest using their Play CDN in production.
The Makefile is also extended to build the
Tailwind styles automatically along with the
HTML generation scripts so the Tailwind styles
should thus not create any overhead for anyone
just updating the text content of the docs.

[STORM-460]